Splitting out smart_holder_init_inline_include.h.
